feat: :sparkles: Se creó el servicio para géneros, un archivo de importación para insertar los géneros y se completó la funcionalidad de guardar y listar actores en la base de datos

+++ src/main/java/com/cine/actor/DTO/DetailActor.java
@@ -7,7 +7,7 @@ public record DetailActor(
     String name,
     LocalDate birthday,
     String profile_path,
-    int gender,
+    Long gender,
     String place_of_birth,
     Double popularity,
     String biography) {}
